Two Optimizations that fix mysql running forever 1. Skiping redundent functions in inter-procedural data flow analysis For each instruction, our pass checks if the operators are contained within the corrupted variable sets. If it is, then put this instruction into the corrutped variable. As a result, to a function, our analysis is not idempotent. We can not simply scan each function once. A simple example is following:
global c
func(a, b<-corrupted)
{

f = c;

c = b;

}

If our pass scan func twice on two different call sites, in the first time,
the result will be {c} and on the second time, the result will be {c, f}.

There is a way to skip functions in a complete and sound way. Right now, for
each function, we keep two values. The first value is the size of the corrupted
variable set when going into the function. The second value is the size of the
corrupted variable set when returning from the function. On a specific call site,
we go into this function based on two conditions.
1) This function has never been analyzed before.
or
2) The current size of the corrupted variable set at the call site is bigger than the enter value
of this function(again, the enter value means the size of the corrupted variable set when going
into this particular function at the most recent call site).

This two conditions gurantees a closure. If the current size is equal to the
enter value of the function. It means that during the last analysis of this function,
no new corrupted variable has been added and on this current call site, the corrupted
variable set is still the same, which means there still won't be any newly added value
when going into this function again. Thus, we can safely skip this callsite.

So, for this case, when func is first analyzed, the enter value is 0 and exit value is 1.
So it will be analyzed again. After this, on the second call site, the enter value is 1
and the exit value is 2, so it will be guranteed to be analyzed again. If there is third call site,
with the corrupted variable set unchanged, this function won't be analyzed again.

2. Store the dominators instead of calculating them again and again for each function.
3. For apache, adjust the test case, because I've updated the bc file containing apr & apr_util.